refactor(virgo): extract events into event service
