+++ packages/global/src/types.ts
@@ -50,8 +50,4 @@ export type BlockModelProps = {
   [K in keyof BlockSchemas]: ReturnType<BlockSchemas[K]['model']['props']>;
 };
 
-export type UnionToIntersection<T> =
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  (T extends any ? (x: T) => any : never) extends (x: infer R) => any
-    ? R
-    : never;
+export * from './utils/types.js';
