+++ packages/virgo/src/utils/index.ts
@@ -1,4 +1,8 @@
 export * from './attributes-renderer.js';
 export * from './base-attributes.js';
 export * from './convert.js';
+export * from './guard.js';
+export * from './query.js';
 export * from './renderer.js';
+export * from './selection.js';
+export * from './text.js';
