+++ packages/virgo/src/utils/text.ts
@@ -1,0 +1,9 @@
+import { ZERO_WIDTH_SPACE } from '../constant.js';
+
+export function calculateTextLength(text: Text): number {
+  if (text.wholeText === ZERO_WIDTH_SPACE) {
+    return 0;
+  } else {
+    return text.wholeText.length;
+  }
+}
